% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_main.R
\name{AStest}
\alias{AStest}
\title{AStest}
\usage{
AStest(y, X, C)
}
\arguments{
\item{y}{phenotype data: n by 1}

\item{X}{genotype data: n by m}

\item{C}{covariated defined by user: n by t}
}
\value{
p values list for each SNP
}
\description{
\code{Astest} is a function to implement GWAS using covariates.
}
\details{
This function use GLM modeling snp, covariates and phenotype.
}
\examples{
P_values <- AStest(y=myY, X=myGD, C=myC)

}
\references{
Lecture 13_GLM in Crops545 Statistical Genomics
}
