% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_main.R
\name{GWAS}
\alias{GWAS}
\title{GWAS}
\usage{
GWAS(y, X, C, pcas = 3)
}
\arguments{
\item{y}{phenotype data: n by 1}

\item{X}{genotype data: n by m}

\item{C}{covariated defined by user: n by t}

\item{pcas}{the number of pcas used in GWAS. Defalut is 3. We can use 'False' to cancel the usage of pcas}
}
\value{
p values list for each SNP
}
\description{
\code{GWAS} is a function to implement GWAS using covariates and PCs from PCA.
}
\details{
This function use GLM modeling snp, covariates, PCs and phenotype.
}
\examples{
# to include 3 PCs in calculation
P_values01 <- GWAS(y=myY, X=myGD, C=myC, pcas=3)
# Dosen't include PCs in calculation
P_values02 <- GWAS(y=myY, X=myGD, C=myC, pcas=False)

}
\references{
Lecture 13_GLM in Crops545 Statistical Genomics
}
